Rename functions and files
